common_drivers/platform_drivers- allow overwrite by customized spi/gpio cores
