+++ common_drivers/platform_drivers/platform_drivers.h
@@ -43,123 +43,20 @@
 /***************************** Include Files **********************************/
 /******************************************************************************/
 
-#ifdef ALTERA
-#include <io.h>
-#include <unistd.h>
-#include <stdio.h>
-#include <system.h>
-#include <alt_stdio.h>
-#include <alt_cache.h>
-#include <alt_types.h>
-#endif
-
-#ifdef NIOS_II
-#endif
-
-#ifdef XILINX
-#include <xparameters.h>
-#include <xil_printf.h>
-#include "platform.h"
-#endif
-
-#ifdef ZYNQ
-#include <sleep.h>
-#include <xspips.h>
-#include <xuartps.h>
-#endif
-
-#ifdef MICROBLAZE
-#include <xil_io.h>
-#include <xil_cache.h>
-#endif
+#include "platform_common.h"
+#include "platform_gpio.h"
+#include "platform_spi.h"
 
 /******************************************************************************/
 /********************** Macros and Constants Definitions **********************/
 /******************************************************************************/
 
-// data types using stdint-
 
-#ifdef ALTERA
-#define int8_t alt_8
-#define int32_t alt_32
-#define uint8_t alt_u8
-#define uint16_t alt_u16
-#define uint32_t alt_u32
-#define uint64_t alt_u64
-#endif
 
-// sleep functions
 
-#define mdelay(msecs) usleep(1000*msecs)
-#define udelay(usecs) usleep(usecs)
 
-#ifdef MICROBLAZE
-void usleep(uint32_t us_count);
-#endif
 
-// print functions
 
-#ifdef ALTERA
-#define ad_printf printf
-#endif
-
-#ifdef XILINX
-#define ad_printf xil_printf
-#endif
-
-// io read/write
-
-#ifdef ALTERA
-#define ad_reg_write(x,y) IOWR_32DIRECT(x,0,y)
-#define ad_reg_read(x) IORD_32DIRECT(x,0)
-#endif
-
-#ifdef XILINX
-#define ad_reg_write(x,y) Xil_Out32(x,y)
-#define ad_reg_read(x) Xil_In32(x)
-#endif
-
-// cache functions
-
-#ifdef ALTERA
-#define ad_icache_flush alt_icache_flush_all
-#define ad_dcache_flush alt_icache_flush_all
-#endif
-
-#ifdef XILINX
-#define ad_icache_flush Xil_ICacheFlush
-#define ad_dcache_flush Xil_DCacheFlush
-#endif
-
-#ifdef ZYNQ
-void Xil_ICacheEnable();
-void Xil_ICacheDisable();
-void Xil_DCacheEnable();
-void Xil_DCacheDisable();
-void Xil_DCacheFlush();
-#endif
-
-// platform functions
-
-#ifdef XILINX
-#define ad_platform_init init_platform
-#define ad_platform_close cleanup_platform
-#endif
-
-#ifdef ALTERA
-void ad_platform_init(void);
-void ad_platform_close(void);
-#endif
-
-// common macros
-
-#define min(x, y) (((x) < (y)) ? (x) : (y))
-#define min_t(type, x, y) (type)min((type)(x), (type)(y))
-#define max(x, y) (((x) > (y)) ? (x) : (y))
-#define max_t(type, x, y) (type)max((type)(x), (type)(y))
-#define clamp(val, min_val, max_val) (max(min((val), (max_val)), (min_val)))
-#define clamp_t(type, val, min_val, max_val) (type)clamp((type)(val), \
-		(type)(min_val), (type)(max_val))
 
 /******************************************************************************/
 /********************** SPI structure and functions ***************************/
@@ -168,33 +65,15 @@ void ad_platform_close(void);
 // you may override the default spi structure-- the drivers simply turn around
 // the spi structure back to spi function calls.
 
-typedef struct {
-	uint32_t    base_address;
-	uint32_t    device_id;
-	uint8_t     chip_select;
-	uint32_t    cpha;
-	uint32_t    cpol;
-} spi_device;
-
-int32_t ad_spi_init(spi_device *dev);
-int32_t ad_spi_xfer(spi_device *dev, uint8_t *data, uint8_t no_of_bytes);
-
 /******************************************************************************/
 /********************* GPIO structure and functions ***************************/
 /******************************************************************************/
 
-int32_t ad_gpio_set(uint8_t pin, uint8_t data);
-int32_t ad_gpio_get(uint8_t pin, uint8_t *data);
-int32_t ad_gpio_set_range(uint8_t start_pin, uint8_t num_pins, uint8_t data);
-int32_t ad_gpio_get_range(uint8_t start_pin, uint8_t num_pins, uint32_t *data);
 
 /******************************************************************************/
 /********************* MISC structure and functions ***************************/
 /******************************************************************************/
 
-uint64_t do_div(uint64_t* n, uint64_t base);
-void ad_reg_write_16(uint32_t addr, uint32_t data);
 uint8_t ad_uart_read();
-uint32_t ad_pow2(uint32_t number);
 
 #endif
